XMLHttpRequest/tests/submissions/Opera/send-redirect-bogus.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - Redirects (bogus Location header)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" data-tested-assertations="/following::dfn[@id='same-origin-request-steps']/DL[1]/DD[2]/P[3]/A[3]" />  
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#same-origin-request-event-rules" data-tested-assertations="/following::DL[1]/DD[1] " />  
  </head>
  <body>
    <div id="log"></div>
    <script>
      setup({timeout: 21000})

      function redirect(code, location) {
        var test = async_test(document.title + " (" + code + ": " + location + ")", {timeout: 20000})
        test.step(function() {
          var client = new XMLHttpRequest()
          client.onreadystatechange = function() {
            test.step(function() {
              if(client.readyState == 4) {
                assert_equals(client.status, 0)
                assert_equals(client.statusText, "")
                test.done()
              }
            })
          }
          client.open("GET", "resources/redirect.php?location=" + location + "&code=" + code)
          client.send(null)
        })
      }
      redirect("302", "http://example.not")
      redirect("302", "mailto:someone@example.org")
      redirect("303", "http://example.not")
      redirect("303", "mailto:someone@example.org")
    </script>
  </body>
</html>
